BTL investors re-think every month if they still want to sell their...
property.
The changes include a separate step in the manageHouse() method, calling the decideToSellHouse() method when household is BTL investor and the property has been on the market. When they decide to not sell the property, they take it off the market.
In the decideToSellInvestmentProperty() method BTL investors do not consider the fact anymore that the house is on the rental market, as it can be on both markets at the same time. To calculate the current rental yield, when the house is also on the rental market there is no rentalRecord price available, which is why I use the avFlowYieldForQuality().